+++ tests/derive/fail/unsupported_option_type.rs
@@ -3,7 +3,7 @@ use kumo_derive::Extract;
 #[derive(Extract)]
 struct Product {
     #[extract(css = ".stock")]
-    stock: Option<u32>,
+    stock: Option<std::time::Duration>,
 }
 
 fn main() {}